requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
cloudinary==1.40.0
coreapi==2.3.3
coreschema==0.0.4
Django==5.2.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
django-jet-reboot==1.3.7
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pyotp==2.9.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.3
typing_extensions==4.11.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.1
